setter for Palette and organize file structure
